requirements.txt
asgiref==3.9.1
Django==5.0
mysqlclient==2.2.7
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
gunicorn==21.2.0
